Animator changed
connection in Animator does not hold functions, it now holds an animatorvariable
this class has a float or bool and an equation. When setbool/setfloat is called in animator it calls doEquation and depending on the class variable it does an equation, returning true when succes. EUI is now on the agenda